Delete core/tee/tee_fs_private.h
Move function prototypes into tee_fs_common.c, tee_rpmb_fs_common.c and
tee_rpmb_fs.c. Next step will be to make them static, remove the
useless wrappers and choose more consistent names.

Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM)
